requirements.txt
yagmail==0.14.247
PyYAML>=5.4
python-dotenv~=0.17.1

requests~=2.26.0